[XDCDecoder] Fix vector type

+++ user/hidra/dc/include/HidraXdcEvent.hh
@@ -9,5 +9,5 @@ struct HidraXdcEvent {
   std::vector<double> ADCflags;
   std::vector<double> TDCvalues;
   std::vector<double> TDCflags;
-  std::vector<uint64_t> XDCTriggers;
+  std::vector<double> XDCTriggers;
 };
